Remove list_head of all nodes.
