+++ src/smc.h
@@ -7,109 +7,6 @@
 #include "tsindex.h"
 #include "evindex.h"
 
-#define NODE_COAL	0
-#define NODE_MIGR	1
-#define NODE_XOVER	2
-#define NODE_SAM	3
-#define NODE_FLOAT	4
-#define NODE_DUMMY	5
-
-#define AS_COAL_NODE(n)	((struct coal_node *)(n))
-#define AS_MIGR_NODE(n)	((struct migr_node *)(n))
-#define AS_XOVER_NODE(n)	((struct xover_node *)(n))
-#define AS_SAM_NODE(n)	((struct sam_node *)(n))
-#define AS_FLOAT_NODE(n)	((struct dummy_node *)(n))
-#define AS_DUMMY_NODE(n)	((struct dummy_node *)(n))
-
-static inline int iscoalnode(struct node *nd)
-{
-	return nd->type == NODE_COAL;
-}
-
-static inline int isxovernode(struct node *nd)
-{
-	return nd->type == NODE_XOVER;
-}
-
-static inline int ismigrnode(struct node *nd)
-{
-	return nd->type == NODE_MIGR;
-}
-
-static inline int issamnode(struct node *nd)
-{
-	return nd->type == NODE_SAM;
-}
-
-static inline int isfloatnode(struct node *nd)
-{
-	return nd->type == NODE_FLOAT;
-}
-
-static inline int isdummynode(struct node *nd)
-{
-	return nd->type == NODE_DUMMY;
-}
-
-static int node_flag_gettype(struct node *nd)
-{
-	return nd->type;
-}
-
-static void edge_flag_settype(struct node *nd, char type)
-{
-	nd->type = type;
-}
-
-static inline char isdeleted(struct node *e)
-{
-	return e->deleted;
-}
-
-static inline void edge_flag_delete(struct node *e)
-{
-	e->deleted = 1;
-}
-
-static inline void edge_flag_undelete(struct node *e)
-{
-	e->deleted = 0;
-}
-
-static inline void edge_flag_setdeleted(struct node *e, char flag)
-{
-	e->deleted = flag;
-}
-
-static inline int edge_flag_getitop(struct node *e)
-{
-	return e->itop;
-}
-
-static inline void edge_flag_setleft(struct node *e)
-{
-	e->itop = 0;
-}
-
-static inline void edge_flag_setright(struct node *e)
-{
-	e->itop = 1;
-}
-
-static inline void edge_flag_setitop(struct node *e, char flag)
-{
-	e->itop = flag;
-}
-
-void free_node(struct genealogy *G, struct node *nd);
-struct node *alloc_node(struct genealogy *G, int type, int pop, double t);
-
-struct node_set {
-	int maxn;
-	int n;
-	struct node **nodes;
-};
-
 struct population {
 	int n;
 	double grate;			// Growth rate at tlast
